Minor changes per PR comments
