+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy.test/src/com/google/cloud/tools/eclipse/appengine/deploy/AppEngineDeployOutputTest.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2016 Google Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.google.cloud.tools.eclipse.appengine.deploy;
 
 import com.google.gson.JsonParseException;
@@ -5,9 +21,9 @@ import org.junit.Assert;
 import org.junit.Test;
 
 /**
- * Unit tests for {@link AppEngineDeployUtil}
+ * Unit tests for {@link AppEngineDeployOutput}
  */
-public class AppEngineDeployUtilTest {
+public class AppEngineDeployOutputTest {
   @Test
   public void testDeployOutputJsonParsingOneVersion() {
     String jsonOutput =
@@ -25,8 +41,8 @@ public class AppEngineDeployUtilTest {
             "  ]\n" +
             "}\n";
 
-    AppEngineDeployUtil.DeployOutput deployOutput =
-        AppEngineDeployUtil.parseDeployOutput(jsonOutput);
+    AppEngineDeployOutput deployOutput =
+        AppEngineDeployOutput.parseDeployOutput(jsonOutput);
     Assert.assertEquals(deployOutput.getVersion(), "20160429t112518");
     Assert.assertEquals(deployOutput.getService(), "default");
   }
@@ -57,7 +73,7 @@ public class AppEngineDeployUtilTest {
             "}\n";
 
     try {
-      AppEngineDeployUtil.parseDeployOutput(jsonOutput);
+      AppEngineDeployOutput.parseDeployOutput(jsonOutput);
       Assert.fail();
     } catch (JsonParseException e) {
       // Success! Should throw a JsonParseException.
@@ -72,7 +88,7 @@ public class AppEngineDeployUtilTest {
             "}\n";
 
     try {
-      AppEngineDeployUtil.parseDeployOutput(jsonOutput);
+      AppEngineDeployOutput.parseDeployOutput(jsonOutput);
       Assert.fail();
     } catch (JsonParseException e) {
       // Success! Should throw a JsonParseException.
