+++ plugins/com.google.cloud.tools.eclipse.appengine.deploy/META-INF/MANIFEST.MF
@@ -10,17 +10,17 @@ Bundle-ActivationPolicy: lazy
 Require-Bundle: com.google.cloud.tools.appengine;bundle-version="0.2.5",
  com.google.cloud.tools.eclipse.util,
  com.google.guava;bundle-version="20.0.0",
+ com.google.gson,
  org.eclipse.core.commands,
  org.eclipse.core.jobs,
  org.eclipse.core.resources,
  org.eclipse.equinox.common,
  org.eclipse.jst.j2ee.ui,
+ org.eclipse.swt,
+ org.eclipse.ui.workbench,
  org.eclipse.wst.common.modulecore,
  org.eclipse.wst.common.project.facet.core,
- org.eclipse.wst.server.core,
- com.google.gson,
- org.eclipse.ui.workbench,
- org.eclipse.swt
+ org.eclipse.wst.server.core
 Export-Package: com.google.cloud.tools.eclipse.appengine.deploy,
  com.google.cloud.tools.eclipse.appengine.deploy.flex,
  com.google.cloud.tools.eclipse.appengine.deploy.standard
